Migrate diff to a CLI
